implement create endpoint for creating a category
