+++ product-manager/src/main/java/ch/bbw/ap/shop/productmanager/controllers/CategoryController.java
@@ -1,11 +1,10 @@
 package ch.bbw.ap.shop.productmanager.controllers;
 
 import ch.bbw.ap.shop.productmanager.models.Category;
+import ch.bbw.ap.shop.productmanager.models.CategoryRequest;
 import ch.bbw.ap.shop.productmanager.services.CategoryService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
@@ -19,5 +18,10 @@ public class CategoryController {
     @GetMapping
     public List<Category> getCategories() {
         return categoryService.getAll();
+    }
+
+    @PostMapping
+    public Category createCategory(@RequestBody CategoryRequest categoryRequest) {
+        return categoryService.createCategory(categoryRequest);
     }
 }
