+++ product-manager/src/main/java/ch/bbw/ap/shop/productmanager/controllers/ProductController.java
@@ -2,6 +2,7 @@ package ch.bbw.ap.shop.productmanager.controllers;
 
 import ch.bbw.ap.shop.productmanager.models.Category;
 import ch.bbw.ap.shop.productmanager.models.Product;
+import ch.bbw.ap.shop.productmanager.models.ProductRequest;
 import ch.bbw.ap.shop.productmanager.services.ProductService;
 import jakarta.ws.rs.Path;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -31,6 +32,33 @@ public class ProductController {
             return ResponseEntity.notFound().build();
         }
         return ResponseEntity.ok(productService.getById(id));
+    }
+
+    @PostMapping("/products")
+    public Product createProduct(@RequestBody ProductRequest product) {
+        return productService.createProduct(product);
+    }
+
+    @PutMapping("/products/{id}")
+    public ResponseEntity<Product> editProduct(@PathVariable Long id, @RequestBody ProductRequest productRequest) {
+
+        Product product = productService.editProduct(id, productRequest);
+
+        if(product == null) {
+            return ResponseEntity.notFound().build();
+        }
+        return ResponseEntity.ok(product);
+    }
+
+    @DeleteMapping("/products/{id}")
+    public ResponseEntity deleteProduct(@PathVariable Long id) {
+        boolean deleted = productService.deleteProduct(id);
+
+        if(!deleted) {
+            return ResponseEntity.notFound().build();
+        }
+
+        return ResponseEntity.noContent().build();
     }
 
 }
