+++ product-manager/src/main/java/ch/bbw/ap/shop/productmanager/models/CategoryRequest.java
@@ -1,0 +1,13 @@
+package ch.bbw.ap.shop.productmanager.models;
+
+public class CategoryRequest {
+    private String category;
+
+    public String getCategory() {
+        return category;
+    }
+
+    public void setCategory(String category) {
+        this.category = category;
+    }
+}
