+++ product-manager/src/main/java/ch/bbw/ap/shop/productmanager/services/CategoryService.java
@@ -1,6 +1,7 @@
 package ch.bbw.ap.shop.productmanager.services;
 
 import ch.bbw.ap.shop.productmanager.models.Category;
+import ch.bbw.ap.shop.productmanager.models.CategoryRequest;
 
 import java.util.List;
 
@@ -9,4 +10,6 @@ public interface CategoryService {
     List<Category> getAll();
 
     Category getById(Long id);
+
+    Category createCategory(CategoryRequest categoryRequest);
 }
